Support ETags for /endpoints/{id}

+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher/src/main/java/org/wso2/carbon/apimgt/rest/api/publisher/impl/EndpointsApiServiceImpl.java
@@ -1,5 +1,6 @@
 package org.wso2.carbon.apimgt.rest.api.publisher.impl;
 
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.carbon.apimgt.core.api.APIPublisher;
@@ -8,6 +9,7 @@ import org.wso2.carbon.apimgt.core.exception.ExceptionCodes;
 import org.wso2.carbon.apimgt.core.models.Endpoint;
 import org.wso2.carbon.apimgt.core.util.APIMgtConstants;
 import org.wso2.carbon.apimgt.rest.api.common.dto.ErrorDTO;
+import org.wso2.carbon.apimgt.rest.api.common.util.ETagGenerator;
 import org.wso2.carbon.apimgt.rest.api.common.util.RestApiUtil;
 import org.wso2.carbon.apimgt.rest.api.publisher.EndpointsApiService;
 import org.wso2.carbon.apimgt.rest.api.publisher.NotFoundException;
@@ -35,6 +37,12 @@ public class EndpointsApiServiceImpl extends EndpointsApiService {
         String username = "";
         try {
             APIPublisher apiPublisher = RestAPIPublisherUtil.getApiPublisher(username);
+            String existingFingerprint = endpointsEndpointIdGetFingerprint(endpointId, null, null, null, minorVersion);
+            if (!StringUtils.isEmpty(ifMatch) && !StringUtils.isEmpty(existingFingerprint) && !ifMatch
+                    .contains(existingFingerprint)) {
+                return Response.status(Response.Status.PRECONDITION_FAILED).build();
+            }
+
             apiPublisher.deleteEndpoint(endpointId);
             return Response.ok().build();
         } catch (APIManagementException e) {
@@ -51,24 +59,30 @@ public class EndpointsApiServiceImpl extends EndpointsApiService {
     @Override
     public Response endpointsEndpointIdGet(String endpointId
             , String contentType
-            , String ifMatch
-            , String ifUnmodifiedSince
+            , String ifNoneMatch
+            , String ifModifiedSince
             , String minorVersion
     ) throws NotFoundException {
         String username = "";
         try {
             APIPublisher apiPublisher = RestAPIPublisherUtil.getApiPublisher(username);
             Endpoint endPoint = apiPublisher.getEndpoint(endpointId);
-            if (endPoint != null) {
-                return Response.ok().entity(MappingUtil.toEndPointDTO(endPoint)).build();
-            } else {
+            if (endPoint == null) {
                 String msg = "Endpoint not found " + endpointId;
                 log.error(msg);
                 ErrorDTO errorDTO = RestApiUtil.getErrorDTO(ExceptionCodes.ENDPOINT_NOT_FOUND);
-                log.error(msg);
                 return Response.status(ExceptionCodes.ENDPOINT_NOT_FOUND.getHttpStatusCode()).entity(errorDTO).build();
             }
 
+            String existingFingerprint = endpointsEndpointIdGetFingerprint(endpointId, contentType, ifNoneMatch,
+                    ifModifiedSince, minorVersion);
+            if (!StringUtils.isEmpty(ifNoneMatch) && !StringUtils.isEmpty(existingFingerprint) && ifNoneMatch
+                    .contains(existingFingerprint)) {
+                return Response.notModified().build();
+            }
+
+            return Response.ok().entity(MappingUtil.toEndPointDTO(endPoint))
+                    .header("Etag", "\"" + existingFingerprint + "\"").build();
         } catch (APIManagementException e) {
             String errorMessage = "Error while get  Endpoint : " + endpointId;
             HashMap<String, String> paramList = new HashMap<String, String>();
@@ -77,6 +91,22 @@ public class EndpointsApiServiceImpl extends EndpointsApiService {
                     (), paramList);
             log.error(errorMessage, e);
             return Response.status(e.getErrorHandler().getHttpStatusCode()).entity(errorDTO).build();
+        }
+    }
+
+    public String endpointsEndpointIdGetFingerprint(String endpointId, String contentType, String ifMatch,
+            String ifUnmodifiedSince, String minorVersion) {
+        String username = RestApiUtil.getLoggedInUsername();
+        try {
+            String lastUpdatedTime = RestAPIPublisherUtil.getApiPublisher(username).getLastUpdatedTimeOfEndpoint(
+                    endpointId);
+            return ETagGenerator.getETag(lastUpdatedTime);
+        } catch (APIManagementException e) {
+            //gives a warning and let it continue the execution
+            String errorMessage =
+                    "Error while retrieving last updated time of endpoint " + endpointId;
+            log.error(errorMessage, e);
+            return null;
         }
     }
 
@@ -93,20 +123,27 @@ public class EndpointsApiServiceImpl extends EndpointsApiService {
             APIPublisher apiPublisher = RestAPIPublisherUtil.getApiPublisher(username);
             Endpoint endpoint = MappingUtil.toEndpoint(body);
             Endpoint retrievedEndpoint = apiPublisher.getEndpoint(endpointId);
-            if (retrievedEndpoint != null) {
-                Endpoint updatedEndpint = new Endpoint.Builder(endpoint).id(endpointId).build();
-                apiPublisher.updateEndpoint(updatedEndpint);
-                Endpoint updatedEndpoint = apiPublisher.getEndpoint(endpointId);
-                return Response.ok().entity(MappingUtil.toEndPointDTO(updatedEndpoint)).build();
-            } else {
+            if (retrievedEndpoint == null) {
                 String msg = "Endpoint not found " + endpointId;
                 log.error(msg);
                 ErrorDTO errorDTO = RestApiUtil.getErrorDTO(ExceptionCodes.ENDPOINT_NOT_FOUND);
-                log.error(msg);
                 return Response.status(ExceptionCodes.ENDPOINT_NOT_FOUND.getHttpStatusCode()).entity(errorDTO).build();
             }
+
+            String existingFingerprint = endpointsEndpointIdGetFingerprint(endpointId, null, null, null, minorVersion);
+            if (!StringUtils.isEmpty(ifMatch) && !StringUtils.isEmpty(existingFingerprint) && !ifMatch
+                    .contains(existingFingerprint)) {
+                return Response.status(Response.Status.PRECONDITION_FAILED).build();
+            }
+
+            Endpoint updatedEndpint = new Endpoint.Builder(endpoint).id(endpointId).build();
+            apiPublisher.updateEndpoint(updatedEndpint);
+            Endpoint updatedEndpoint = apiPublisher.getEndpoint(endpointId);
+            String newFingerprint = endpointsEndpointIdGetFingerprint(endpointId, null, null, null, minorVersion);
+            return Response.ok().header("Etag", "\"" + newFingerprint + "\"")
+                    .entity(MappingUtil.toEndPointDTO(updatedEndpoint)).build();
         } catch (APIManagementException e) {
-            String errorMessage = "Error while get All Endpoint";
+            String errorMessage = "Error while getting the endpoint :" + endpointId;
             org.wso2.carbon.apimgt.rest.api.common.dto.ErrorDTO errorDTO = RestApiUtil.getErrorDTO(e.getErrorHandler());
             log.error(errorMessage, e);
             return Response.status(e.getErrorHandler().getHttpStatusCode()).entity(errorDTO).build();
